backend/requirements.txt
aniso8601==8.0.0
async-generator==1.10
attrs==22.1.0
cachetools==5.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==7.1.2
et-xmlfile==1.1.0
exceptiongroup==1.0.0rc9
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
google-api-core==2.10.1
google-api-python-client==2.63.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-cloud-vision==3.1.3
googleapis-common-protos==1.56.4
greenlet==1.1.3
grpcio==1.49.1
grpcio-status==1.49.1
gunicorn==20.1.0
h11==0.14.0
httplib2==0.20.4
idna==3.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.23.3
openpyxl==3.0.10
outcome==1.2.0
pandas==1.5.0
proto-plus==1.22.1
protobuf==4.21.7
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2020.1
requests==2.28.1
rsa==4.9
selenium==4.5.0
six==1.15.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.16
trio==0.22.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.12
Werkzeug==1.0.1
wsproto==1.2.0
xlrd==2.0.1
